+++ examples/react-router-tld/app/po.d.ts
@@ -1,7 +1,7 @@
 declare module "*.po" {
-  export const messages: Record<string, string>
+  export const messages: import("@palamedes/core/compiled").CompiledCatalogMessages
   const catalog: {
-    messages: Record<string, string>
+    messages: import("@palamedes/core/compiled").CompiledCatalogMessages
   }
   export default catalog
 }